+++ 1.20/Common/src/main/java/fuzs/blockrunner/world/level/block/data/BlockSpeedManager.java
@@ -15,27 +15,29 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.server.packs.resources.PreparableReloadListener;
 import net.minecraft.server.packs.resources.ResourceManager;
+import net.minecraft.server.packs.resources.ResourceManagerReloadListener;
+import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.TagKey;
-import net.minecraft.util.Unit;
-import net.minecraft.util.profiling.ProfilerFiller;
+import net.minecraft.util.GsonHelper;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
 
 import java.io.File;
 import java.io.FileReader;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.Executor;
 
-public class BlockSpeedManager implements PreparableReloadListener {
+public class BlockSpeedManager implements ResourceManagerReloadListener {
     public static final BlockSpeedManager INSTANCE = new BlockSpeedManager();
+    public static final String SCHEMA_VERSION = String.valueOf(2);
     public static final UUID SPEED_MODIFIER_BLOCK_SPEED_UUID = UUID.fromString("23237052-61AD-11EB-AE93-0242AC130002");
     private static final String CONFIG_FILE_NAME = BlockRunner.MOD_ID + ".json";
     private static final Set<SpeedHolderValue> DEFAULT_BLOCK_SPEEDS = new SpeedHolderValue.Builder()
+            .add(BlockTags.STONE_BRICKS, 1.15)
+            .add(Blocks.DIRT_PATH, 1.35)
             .add(ModRegistry.VERY_SLOW_BLOCKS_BLOCK_TAG, 0.45)
             .add(ModRegistry.SLOW_BLOCKS_BLOCK_TAG, 0.65)
             .add(ModRegistry.SLIGHTLY_SLOW_BLOCKS_BLOCK_TAG, 0.85)
@@ -54,11 +56,7 @@ public class BlockSpeedManager implements PreparableReloadListener {
     }
 
     @Override
-    public final CompletableFuture<Void> reload(PreparableReloadListener.PreparationBarrier p_10780_, ResourceManager p_10781_, ProfilerFiller p_10782_, ProfilerFiller p_10783_, Executor p_10784_, Executor p_10785_) {
-        return p_10780_.wait(Unit.INSTANCE).thenRunAsync(this::load, p_10785_);
-    }
-
-    public void load() {
+    public void onResourceManagerReload(ResourceManager resourceManager) {
         JsonConfigFileUtil.getAndLoad(CONFIG_FILE_NAME, this::serialize, this::deserialize);
         // this is also called when the server is starting and Forge has not set the game server object yet
         if (ModLoaderEnvironment.INSTANCE.isServer() && Proxy.INSTANCE.getGameServer() != null) {
@@ -97,6 +95,7 @@ public class BlockSpeedManager implements PreparableReloadListener {
 
     private JsonObject serialize(Set<SpeedHolderValue> values) {
         JsonObject jsonElements = new JsonObject();
+        jsonElements.addProperty("schemaVersion", SCHEMA_VERSION);
         for (SpeedHolderValue value : values) {
             value.serialize(jsonElements);
         }
@@ -107,26 +106,32 @@ public class BlockSpeedManager implements PreparableReloadListener {
         this.deserialize(JsonConfigFileUtil.GSON.fromJson(reader, JsonObject.class));
     }
 
-    public void deserialize(JsonObject jsonElements) {
+    public void deserialize(JsonObject jsonObject) {
         this.blockSpeedValues.clear();
         this.blockSpeeds = null;
-        Set<SpeedHolderValue> blockSpeedValues = Sets.newHashSet();
-        for (Map.Entry<String, JsonElement> entry : jsonElements.entrySet()) {
+        Map<Object, SpeedHolderValue> blockSpeedValues = Maps.newIdentityHashMap();
+        String schemaVersion = GsonHelper.getAsString(jsonObject, "schema_version", "1");
+        if (!schemaVersion.equals(SCHEMA_VERSION)) {
+            BlockRunner.LOGGER.warn("Outdated config schema! Config might not work correctly. Current schema is {}.", SCHEMA_VERSION);
+            blockSpeedValues.put(BlockTags.STONE_BRICKS, new SpeedHolderValue.TagValue(BlockTags.STONE_BRICKS, 1.15));
+            blockSpeedValues.put(Blocks.DIRT_PATH, new SpeedHolderValue.BlockValue(Blocks.DIRT_PATH, 1.35));
+        }
+        for (Map.Entry<String, JsonElement> entry : jsonObject.entrySet()) {
             String key = entry.getKey();
             double speedValue = entry.getValue().getAsDouble();
             if (key.startsWith("#")) {
                 TagKey<Block> tag = TagKey.create(Registries.BLOCK, new ResourceLocation(key.substring(1)));
-                blockSpeedValues.add(new SpeedHolderValue.TagValue(tag, speedValue));
+                blockSpeedValues.put(tag, new SpeedHolderValue.TagValue(tag, speedValue));
             } else {
                 ResourceLocation resourcelocation = new ResourceLocation(key);
                 if (BuiltInRegistries.BLOCK.containsKey(resourcelocation)) {
                     Block block = BuiltInRegistries.BLOCK.get(resourcelocation);
-                    blockSpeedValues.add(new SpeedHolderValue.BlockValue(block, speedValue));
+                    blockSpeedValues.put(block, new SpeedHolderValue.BlockValue(block, speedValue));
                 } else {
                     BlockRunner.LOGGER.warn("Unknown block type '{}', valid types are: {}", resourcelocation, Joiner.on(", ").join(BuiltInRegistries.BLOCK.keySet()));
                 }
             }
         }
-        this.blockSpeedValues.addAll(blockSpeedValues);
+        this.blockSpeedValues.addAll(blockSpeedValues.values());
     }
 }