+++ 1.20/Forge/src/main/java/fuzs/blockrunner/BlockRunnerForge.java
@@ -1,18 +1,12 @@
 package fuzs.blockrunner;
 
 import fuzs.blockrunner.data.ModBlockTagsProvider;
-import fuzs.blockrunner.data.ModLanguageProvider;
+import fuzs.blockrunner.data.client.ModLanguageProvider;
 import fuzs.puzzleslib.api.core.v1.ModConstructor;
-import net.minecraft.core.HolderLookup;
-import net.minecraft.data.DataGenerator;
-import net.minecraft.data.PackOutput;
-import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.data.event.GatherDataEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
-
-import java.util.concurrent.CompletableFuture;
 
 @Mod(BlockRunner.MOD_ID)
 @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.MOD)
@@ -25,11 +19,7 @@ public class BlockRunnerForge {
 
     @SubscribeEvent
     public static void onGatherData(final GatherDataEvent evt) {
-        final DataGenerator dataGenerator = evt.getGenerator();
-        final PackOutput packOutput = dataGenerator.getPackOutput();
-        final CompletableFuture<HolderLookup.Provider> lookupProvider = evt.getLookupProvider();
-        final ExistingFileHelper fileHelper = evt.getExistingFileHelper();
-        dataGenerator.addProvider(true, new ModBlockTagsProvider(packOutput, lookupProvider, BlockRunner.MOD_ID, fileHelper));
-        dataGenerator.addProvider(true, new ModLanguageProvider(packOutput, BlockRunner.MOD_ID));
+        evt.getGenerator().addProvider(true, new ModBlockTagsProvider(evt, BlockRunner.MOD_ID));
+        evt.getGenerator().addProvider(true, new ModLanguageProvider(evt, BlockRunner.MOD_ID));
     }
 }