+++ 1.20/Forge/src/main/java/fuzs/blockrunner/data/ModBlockTagsProvider.java
@@ -4,16 +4,15 @@ import fuzs.blockrunner.init.ModRegistry;
 import fuzs.puzzleslib.api.data.v1.AbstractTagProvider;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
-import net.minecraft.tags.BlockTags;
-import net.minecraft.world.level.block.Blocks;
 import net.minecraftforge.common.data.ExistingFileHelper;
+import net.minecraftforge.data.event.GatherDataEvent;
 
 import java.util.concurrent.CompletableFuture;
 
 public class ModBlockTagsProvider extends AbstractTagProvider.Blocks {
 
-    public ModBlockTagsProvider(PackOutput packOutput, CompletableFuture<HolderLookup.Provider> lookupProvider, String modId, ExistingFileHelper fileHelper) {
-        super(packOutput, lookupProvider, modId, fileHelper);
+    public ModBlockTagsProvider(GatherDataEvent evt, String modId) {
+        super(evt, modId);
     }
 
     @Override
@@ -21,8 +20,8 @@ public class ModBlockTagsProvider extends AbstractTagProvider.Blocks {
         this.tag(ModRegistry.VERY_SLOW_BLOCKS_BLOCK_TAG);
         this.tag(ModRegistry.SLOW_BLOCKS_BLOCK_TAG);
         this.tag(ModRegistry.SLIGHTLY_SLOW_BLOCKS_BLOCK_TAG);
-        this.tag(ModRegistry.SLIGHTLY_QUICK_BLOCKS_BLOCK_TAG).addTag(BlockTags.STONE_BRICKS);
-        this.tag(ModRegistry.QUICK_BLOCKS_BLOCK_TAG).add(Blocks.DIRT_PATH);
+        this.tag(ModRegistry.SLIGHTLY_QUICK_BLOCKS_BLOCK_TAG);
+        this.tag(ModRegistry.QUICK_BLOCKS_BLOCK_TAG);
         this.tag(ModRegistry.VERY_QUICK_BLOCKS_BLOCK_TAG);
     }
 }