+++ 1.20/Forge/src/main/java/fuzs/blockrunner/data/client/ModLanguageProvider.java
@@ -1,12 +1,13 @@
-package fuzs.blockrunner.data;
+package fuzs.blockrunner.data.client;
 
 import fuzs.puzzleslib.api.data.v1.AbstractLanguageProvider;
 import net.minecraft.data.PackOutput;
+import net.minecraftforge.data.event.GatherDataEvent;
 
 public class ModLanguageProvider extends AbstractLanguageProvider {
 
-    public ModLanguageProvider(PackOutput packOutput, String modId) {
-        super(packOutput, modId);
+    public ModLanguageProvider(GatherDataEvent evt, String modId) {
+        super(evt, modId);
     }
 
     @Override