+++ 1.20/CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v8.0.1-1.20.1] - 2023-11-06
+### Changed
+- Moved default block and block tag configurations to `blockrunner.json` instead of adding those to the tags themselves added by Block Runner
+- This way users may change default values without the need for a data pack
+
 ## [v8.0.0-1.20.1] - 2023-06-27
 - Ported to Minecraft 1.20.1